moved PluginList to plugins.js
