+++ src/segment-parsers/tiff-exif.js
@@ -1,4 +1,5 @@
-import {AppSegmentParserBase, segmentParsers} from '../parser.js'
+import {AppSegmentParserBase} from '../parser.js'
+import {segmentParsers} from '../plugins.js'
 import {TAG_IFD_EXIF, TAG_IFD_GPS, TAG_IFD_INTEROP, TAG_MAKERNOTE, TAG_USERCOMMENT, TAG_XMP, TAG_IPTC, TAG_ICC} from '../tags.js'
 import {TAG_GPS_LATREF, TAG_GPS_LAT, TAG_GPS_LONREF, TAG_GPS_LON} from '../tags.js'
 import {TIFF_LITTLE_ENDIAN, TIFF_BIG_ENDIAN} from '../util/helpers.js'
